Conflicts:
	nestkernel/connection_creator_impl.h
	pynest/examples/Potjans_2014/network.py
	pynest/examples/brunel_exp_multisynapse_nest.py
	pynest/examples/lin_rate_ipn_network.py
	pynest/examples/store_restore_network.py
	pynest/examples/structural_plasticity.py
	pynest/examples/synapsecollection.py
	pynest/examples/urbanczik_synapse_example.py
	pynest/nest/__init__.py
	pynest/nest/lib/hl_api_connections.py
	testsuite/pytests/test_NodeCollection.py
	testsuite/pytests/test_connect_all_to_all.py
	testsuite/pytests/test_connect_arrays_mpi.py
	testsuite/pytests/test_connect_fixed_indegree.py
	testsuite/pytests/test_connect_fixed_outdegree.py
	testsuite/pytests/test_connect_fixed_total_number.py
	testsuite/pytests/test_connect_helpers.py
	testsuite/pytests/test_connect_one_to_one.py
	testsuite/pytests/test_connect_pairwise_bernoulli.py
	testsuite/pytests/test_connect_parameters.py
	testsuite/pytests/test_connect_symmetric_pairwise_bernoulli.py
	testsuite/pytests/test_glif_cond.py
	testsuite/pytests/test_nodeParametrization.py
	testsuite/pytests/test_sp/test_disconnect.py
	testsuite/pytests/test_sp/test_disconnect_multiple.py
	testsuite/pytests/test_sp/test_sp_manager.py
	testsuite/pytests/test_spatial/test_basics.py
	testsuite/pytests/test_spatial/test_connect_layers.py

+++ nestkernel/mpiutil.h
@@ -38,7 +38,7 @@ namespace shadow
 using cell_node_id_type = std::uint32_t;
 using cell_lid_type = std::uint32_t;
 struct cell_member_type;
-using time_type = float;
+using time_type = double;
 constexpr time_type terminal_time = std::numeric_limits< time_type >::max();
 
 template < typename I >
@@ -51,7 +51,7 @@ int mpi_rank( MPI_Comm c );
 int mpi_size( MPI_Comm c );
 int broadcast( int local, MPI_Comm comm, int root );
 unsigned broadcast( unsigned local, MPI_Comm comm, int root );
-float broadcast( float local, MPI_Comm comm, int root );
+double broadcast( double local, MPI_Comm comm, int root );
 
 #ifdef HAVE_RECORDINGBACKEND_ARBOR
 struct comm_info;
